refactor(apps/client): HASHI-165 token reissue를 shared로 이동
